handshake utilities to separate file
